gmail/gmail-requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.57
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mplcursors==0.5.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.20.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.0
openai==1.30.1
opencv-python==4.9.0.80
orjson==3.10.3
packaging==23.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydyf==0.10.0
pyparsing==3.1.2
pypdfium2==4.30.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-doctr==0.8.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sympy==1.12
tenacity==8.3.0
tiktoken==0.7.0
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typing-inspect==0.9.0
typing_extensions==4.11.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==62.1
webencodings==0.5.1
yarl==1.9.4
zipp==3.18.1
zopfli==0.2.3
flask
python-dotenv
flask-cors